api/pyproject.toml
[build-system]
requires = ['setuptools', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = "mnaccounts"
requires-python = ">=3.7"

#author = "Markiyan Kushnr"
#author-email = "markiyan.kushnir@gmail.com"
description = "Accounts Service"
#url = "https://accouns.mkushir.mooo.com"

dependencies = [
    "flask-restful",
    "flask-cors",
    "flask-login",
    "flask-bcrypt",
    "oauthlib",
    "dateparser",
    "SQLAlchemy",
    "requests",
    "pyjwt[crypto]",
    "gunicorn",
    "alembic",
    "mnaccounts-client @ git+https://github.com/mkushnir/mnaccounts-client.git"
]

dynamic = [
    "version",
]

[tool.setuptools]
packages = [
    "mnaccounts",
    "mnaccounts.model",
    "mnaccounts.resource",
]

[tool.setuptools.dynamic]
version = {attr = "mnaccounts.version.__version__"}
